Used the observor pattern for the Save_Load class and changed the Save_Load class to GameState. I also unembedded the Character class from the xaml class because of code smell (programming lingo).
