+++ Battle of the Professors/Battle of the Professor/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:Battle_of_the_Professor"
-             StartupUri="MainWindow.xaml">
+             StartupUri="MainMenu.xaml">
     <Application.Resources>
          
     </Application.Resources>
