+++ Battle of the Professors/Battle of the Professor/Deprived.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+
+namespace Battle_of_the_Professor
+{
+    public class Deprived : Character
+    {
+        public Deprived(int health, int sanity, int intellect) : base(health, sanity, intellect)
+        {
+
+        }
+    }
+}
