+++ Battle of the Professors/Battle of the Professor/IGameState.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Battle_of_the_Professor
+{
+    public interface IGameState
+    {
+        void Save(Character player);
+    }
+}
